requirements.txt
flask~=3.1.0
pandas~=2.1.4
gunicorn

python-dateutil~=2.9.0.post0
numpy~=1.26.4
altair~=4.2.2
fuzzywuzzy~=0.18.0
sentence-transformers~=3.4.1
python-Levenshtein
